# backend/openapi/v2.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "XUGOU v2 API",
    "version": "2.0.0",
    "description": "Single-Worker modular-monolith API contract."
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "security": [
    {
      "sessionCookie": []
    }
  ],
  "paths": {
...
    "/api/v2/operations/compatibility-hits": {
      "get": {
        "operationId": "listCompatibilityHits",
        "tags": [
          "operations"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/CompatibilityHit"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Problem Details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblem"
                }
              }
            }
          },
          "401": {
            "description": "Problem Details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblem"
                }
              }
            }
          },
          "500": {
            "description": "Problem Details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblem"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "days",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 365,
              "default": 30
            }
          }
        ]
      }
    },
...
    "/api/v2/status/config": {
      "get": {
        "operationId": "getStatusConfigV2",
        "tags": [
          "status"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/StatusConfigView"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Problem Details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblem"
                }
              }
            }
          },
          "500": {
            "description": "Problem Details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblem"
                }
              }
            }
          }
        }
      },
      "put": {
        "operationId": "saveStatusConfigV2",
        "tags": [
          "status"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/StatusConfigView"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Problem Details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblem"
                }
              }
            }
          },
          "401": {
            "description": "Problem Details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblem"
                }
              }
            }
          },
          "403": {
            "description": "Problem Details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblem"
                }
              }
            }
          },
          "500": {
            "description": "Problem Details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblem"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StatusConfigCommand"
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "xugou_session"
      },
      "releaseToken": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "opaque release credential"
      },
      "agentBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    },
    "schemas": {
      "ApiProblem": {
        "type": "object",
        "required": [
          "type",
          "title",
          "status",
          "code",
          "trace_id"
        ],
        "properties": {
          "type": {
            "type": "string",
            "format": "uri"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "integer"
          },
          "code": {
            "type": "string"
          },
          "trace_id": {
            "type": "string"
          },
          "detail": {
            "type": "string"
          },
          "errors": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        }
      },
...
      "CompatibilityHit": {
        "type": "object",
        "required": [
          "day",
          "route_group",
          "method",
          "status_family",
          "hit_count",
          "first_seen_at",
          "last_seen_at",
          "last_release_version"
        ],
        "properties": {
          "day": {
            "type": "string",
            "format": "date"
          },
          "route_group": {
            "type": "string",
            "enum": [
              "monitor_management_v1",
              "agent_management_v1",
              "agent_registration_v1",
              "agent_report_v1",
              "notification_management_v1",
              "status_v1"
            ]
          },
          "method": {
            "type": "string"
          },
          "status_family": {
            "type": "string"
          },
          "hit_count": {
            "type": "integer"
          },
          "first_seen_at": {
            "type": "string",
            "format": "date-time"
          },
          "last_seen_at": {
            "type": "string",
            "format": "date-time"
          },
          "last_release_version": {
            "type": "string"
          }
        }
      },
...
      "StatusConfigCommand": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "title",
          "description",
          "logoUrl",
          "customCss",
          "theme",
          "monitors",
          "agents"
        ],
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 120
          },
          "description": {
            "type": "string",
            "maxLength": 500
          },
          "logoUrl": {
            "type": "string",
            "maxLength": 2048
          },
          "customCss": {
            "type": "string",
            "maxLength": 20000
          },
          "theme": {
            "type": "string"
          },
          "monitors": {
            "type": "array",
            "maxItems": 100,
            "uniqueItems": true,
            "items": {
              "type": "integer"
            }
          },
          "agents": {
            "type": "array",
            "maxItems": 100,
            "uniqueItems": true,
            "items": {
              "type": "integer"
            }
          }
        }
      },
      "StatusConfigView": {
        "type": "object",
        "required": [
          "title",
          "description",
          "logoUrl",
          "customCss",
          "theme",
          "monitors",
          "agents",
          "monitors_has_more",
          "agents_has_more"
        ],
        "properties": {
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "logoUrl": {
            "type": "string"
          },
          "customCss": {
            "type": "string"
          },
          "theme": {
            "type": "string"
          },
          "monitors": {
            "type": "array",
            "maxItems": 500,
            "items": {
              "type": "object",
              "required": [
                "id",
                "name",
                "selected"
              ],
              "properties": {
                "id": {
                  "type": "integer"
                },
                "name": {
                  "type": "string"
                },
                "selected": {
                  "type": "boolean"
                }
              }
            }
          },
          "agents": {
            "type": "array",
            "maxItems": 500,
            "items": {
              "type": "object",
              "required": [
                "id",
                "name",
                "selected"
              ],
              "properties": {
                "id": {
                  "type": "integer"
                },
                "name": {
                  "type": "string"
                },
                "selected": {
                  "type": "boolean"
                }
              }
            }
          },
          "monitors_has_more": {
            "type": "boolean"
          },
          "agents_has_more": {
            "type": "boolean"
          }
        }
      },
...
    }
  }
}